Add provision for the no-battery version
Does an early i2c query of the AXP209 as to whether a battery is present. If a battery is found, the code proceeds as in prior versions. If no battery is found, the OLED display is altered to remove battery info on the first OLED page and remove the battery info page (page 3) completely.