Add branding capability
Added new file "branding.py" which is to be a central repository of information related to the branding of the CB build. (Example: the filename of the splash screen logo and the display name on the main page of the OLED.) Also updated the displays.py and page_main.py files to use the information in branding.py.

It is anticipated that other information related to specific build (ex: pin assignments for unique hardware) can also be added here.