Remove RPi.GPIO install and add 2 python versions
Remove the install of RPi.GPIO 0.7.0 (line 22) as it conflicts with the required RPi.GPIO_NP_CB required for the NanoPi NEO build. The 0.7.0 version is required for RPi versions of the ConnectBox and will be added as a conditional install in the ansible code in file roles/hat-service/tasks/main.yml (similar to the install of RPi.GPIO_NP_CB for the NEO build).
Also added Python 3.7 and 3.8 to the list of acceptable Programming Languages at lines 64 and 65.