Fix GPIO interrupt handler; add globals.py
The RPi.GPIO function GPIO.add_event_detect() makes the interrupt pin unreadable by GPIO.input and thus breaks our need for testing pin state within the interrupt handler. This change adds low level register writes to the interrupt handler in buttons.py to temporarily put make the pins readable as inputs. At the end of the interrupt routine, the interrupt function of the pins are restored.
Additionally, since this need only applies to the NEO processor, a test was added to determine what kind of SBC we are running. This testing was previously found in several places in the code and is now centralized in the new file globals.py.