Fixed possibility of having the USB key mounted on sdb1 or sdc1 etc for /media/usb0 and erase will not erase the USB key it will be removed first.
Note that on copy to /media/usb0 from usb key's you can have one or more keys in place and it will tally the total space and copy all keys present to the main /media/usb0 directory.  Once finished it leaves the keys mounted.  But the automout tool will unmount them when they are removed.