Improve shutdown operation; protect from loss of AXP209 i2c below 3.1V
The AXP209 i2c fails below about 3.1V and can cause a crash of the neo-battery-shutdown code. These changes use the python try/except method to protect the python code when the i2c fails. Changes to cli.py are to make the default logging "INFO" rather than "DEBUG" to make the log easier to read.