Move all ATTiny comm functions to multiBat_Utilities
Handle all i2c communication to ATTiny in multiBat_Utilities.py. Code now correctly handles reading and display of battery voltage when ATTiny communication is down. This also provides for the case where the battery circuitry does not include an ATTiny (ex: if we have batteries hard wired in parallel).